+++ src/internal.ts
@@ -9,6 +9,7 @@ export * from "./utils/utils"
 export * from "./core/atom"
 export * from "./utils/comparer"
 export * from "./utils/decorators"
+export * from "./utils/stage2decorators"
 export * from "./types/modifiers"
 export * from "./api/observabledecorator"
 export * from "./api/observable"
